Aut 3530/reauth lockouts

+++ shared/src/main/java/uk/gov/di/authentication/shared/helpers/ValidationHelper.java
@@ -5,6 +5,7 @@ import com.google.i18n.phonenumbers.PhoneNumberUtil;
 import com.google.i18n.phonenumbers.PhoneNumberUtil.PhoneNumberType;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.jetbrains.annotations.NotNull;
 import uk.gov.di.authentication.shared.entity.ErrorResponse;
 import uk.gov.di.authentication.shared.entity.JourneyType;
 import uk.gov.di.authentication.shared.entity.NotificationType;
@@ -136,7 +137,9 @@ public class ValidationHelper {
             ConfigurationService configurationService) {
 
         if (code.filter(input::equals).isPresent()) {
-            codeStorageService.deleteIncorrectMfaCodeAttemptsCount(emailAddress);
+            if (journeyType != JourneyType.REAUTHENTICATION) {
+                codeStorageService.deleteIncorrectMfaCodeAttemptsCount(emailAddress);
+            }
 
             switch (notificationType) {
                 case MFA_SMS:
@@ -149,25 +152,43 @@ public class ValidationHelper {
             return Optional.of(ErrorResponse.ERROR_1002);
         }
 
-        if (configurationService.supportAccountCreationTTL() && notificationType == VERIFY_EMAIL) {
-            codeStorageService.increaseIncorrectMfaCodeAttemptsCountAccountCreation(emailAddress);
-        } else {
-            codeStorageService.increaseIncorrectMfaCodeAttemptsCount(emailAddress);
-        }
+        return getErrorResponse(
+                notificationType,
+                journeyType,
+                codeStorageService,
+                emailAddress,
+                configurationService);
+    }
 
-        if (codeStorageService.getIncorrectMfaCodeAttemptsCount(emailAddress)
-                > configurationService.getCodeMaxRetries()) {
-            switch (notificationType) {
-                case MFA_SMS:
-                    return Optional.of(ErrorResponse.ERROR_1027);
-                case VERIFY_EMAIL:
-                    return Optional.of(ErrorResponse.ERROR_1033);
-                case VERIFY_CHANGE_HOW_GET_SECURITY_CODES:
-                    return Optional.of(ErrorResponse.ERROR_1048);
-                case VERIFY_PHONE_NUMBER:
-                    return Optional.of(ErrorResponse.ERROR_1034);
-                case RESET_PASSWORD_WITH_CODE:
-                    return Optional.of(ErrorResponse.ERROR_1039);
+    private static @NotNull Optional<ErrorResponse> getErrorResponse(
+            NotificationType notificationType,
+            JourneyType journeyType,
+            CodeStorageService codeStorageService,
+            String emailAddress,
+            ConfigurationService configurationService) {
+        if (journeyType != JourneyType.REAUTHENTICATION) {
+            if (configurationService.supportAccountCreationTTL()
+                    && notificationType == VERIFY_EMAIL) {
+                codeStorageService.increaseIncorrectMfaCodeAttemptsCountAccountCreation(
+                        emailAddress);
+            } else {
+                codeStorageService.increaseIncorrectMfaCodeAttemptsCount(emailAddress);
+            }
+
+            if (codeStorageService.getIncorrectMfaCodeAttemptsCount(emailAddress)
+                    > configurationService.getCodeMaxRetries()) {
+                switch (notificationType) {
+                    case MFA_SMS:
+                        return Optional.of(ErrorResponse.ERROR_1027);
+                    case VERIFY_EMAIL:
+                        return Optional.of(ErrorResponse.ERROR_1033);
+                    case VERIFY_CHANGE_HOW_GET_SECURITY_CODES:
+                        return Optional.of(ErrorResponse.ERROR_1048);
+                    case VERIFY_PHONE_NUMBER:
+                        return Optional.of(ErrorResponse.ERROR_1034);
+                    case RESET_PASSWORD_WITH_CODE:
+                        return Optional.of(ErrorResponse.ERROR_1039);
+                }
             }
         }
